+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.2] - 2026-01-07
+
+### Bug Fixes
+
+- Fix SSL certificate verification for Nuitka binaries on macOS
+- Fix macOS stdlib download URL (use PyPy 3.8 v7.3.11)
+
 ## [0.10.1] - 2026-01-07
 
 ### Enhancements